backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20
httpx[http2]==0.28.1
alembic==1.14.1
aiosqlite==0.20.0
openai>=1.0.0
pyyaml>=6.0
websockets>=12.0
